# openapi.yaml
openapi: 3.0.0
info:
  title: myAirfare CRUD and authorizations
  version: 0.1.0
servers:
- description: Mock Server
  url: http://localhost:3000
...
paths:
...
  /api/v1/update-profile:
    put:
      operationId: updateProfile
      description: Enpoint dengan tags user authorization adalah endpoint yang dapat diakses oleh semua access level mulai dari member sampai admin
      tags:
      - All User Authorization
      parameters:
      - name: x-access-token
        in: header
        required: true
        schema:
          type: string
          format: jwt
          description: Parameter token dengan nama x-access-token adalah parameter wajib untuk mengakses endpoint enpoint pada API ini. x-access-token, bisa didaptakan dari response endpoint login dengan key token
          example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyIjoiNjY1M2YzNjktNGRmNC00M2Y5LWE2NzMtOWVjNjEwMzU5NDI1IiwiaWF0IjoxNjY2MDcwNjA3fQ.a2uU6WZ_wW6zfZOTSYzSLusiuTTczLYCvW7GBfioMvo
      requestBody:
        $ref: "#/components/requestBodies/updateUserRequest"
      responses:
        "200":
          description: "Ini hasil respon edit profil jika request update berhasil"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/update"

        "400":
          description: Ini hasil respon jika proses update profil gagal disimpan ke server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/badRequest"

        "401":
          description: Ini hasil respon error autorisaasi / login gagal
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
                    example: ["Akses dilarang, user tidak terdaftar"]

        "403":
          description: Ini hasil repon error akun user belum aktif / verifikasi atau saat mengubah akun user dengan kondisi nama akun / email sudah ada / terdaftar
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
                    example: ["Akses dilarang,..."]


        "404":
          description: Ini hasil respon error field wajib tidak diisi atau saat konfirmasi password tidak sesuai
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: string
                    example: ["Semua field wajib diisi / Konfirmasi password tidak sesuai"]
...
components:
  schemas:
...
    badRequest:
      type: object
      properties:
        error:
          type: string
          example: ["(server error ...)"]
...
    update:
      type: object
      properties:
        update:
          type: array
          items:
            type: integer
            description: Nilai integer ini adalah jumlah data yang berhasil di update (rows tabel yang terupdate)
            example: 1
...
  requestBodies:
...
    updateUserRequest:
      description: Ini adalah request body data untuk update user / admin. Ubah data profile di request body ini. Semua field update profil, tidak wajib diisi untuk field password, re-type password dan photo.
      content:
        multipart/form-data:
          schema:
            type: object
            required:
            - username
            - l_name
            - f_name
            - email
            - title

            properties:
              username:
                type: string
                example: user
              f_name:
                type: string
                example: user
              l_name:
                type: string
                example: user
              email:
                type: string
                example: user@mail.com
              password:
                type: string
                example: p4ssw0rd
              re_type_pass:
                type: string
                example: p4ssw0rd
              title:
                type: string
                example: Mr
              photo:
                type: string
                format: binary